updated types

+++ common/stats.cpp
@@ -180,7 +180,7 @@ void statmap::period(FILE *fp)
         last = node->lastcall;
         Mutex::release(node);
         if(fp)
-            fprintf(fp, "%s %ld\n", text, last);
+            fprintf(fp, "%s %ld\n", text, (long)last);
     }
 }
